ML_Pro/requirements.txt
abseil-cpp=20211102.0=hd77b12b_0
absl-py=2.1.0=py39haa95532_0
aiohttp=3.10.5=py39h827c3e9_0
aiosignal=1.2.0=pyhd3eb1b0_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
attrs=24.2.0=py39haa95532_0
beautifulsoup4=4.12.3=py39haa95532_0
bleach=6.2.0=pypi_0
brotli=1.0.9=h2bbff1b_8
cudatoolkit=11.2.2=h7d7167e_13
cudnn=8.1.0.77=h3e0f4f4_0
debugpy=1.8.8=pypi_0
fastapi=0.115.6=pypi_0
fastjsonschema=2.20.0=pypi_0
flask=2.3.2=pypi_0
fonttools=4.54.1=pypi_0
google-auth=2.36.0=pypi_0
grpcio=1.67.1=pypi_0
h5py=3.12.1=py39h3b2c811_0
hdf5=1.12.1=h51c971a_3
idna=3.10=pypi_0
ipykernel=6.29.5=py39haa95532_0
ipython=8.18.1=pypi_0
jupyter=1.1.1=pypi_0
keras=2.10.0=py39haa95532_0
matplotlib=3.9.2=pypi_0
numpy=2.0.2=pypi_0
opt-einsum=3.4.0=pypi_0
pillow=11.0.0=pypi_0
protobuf=3.20.0=pypi_0
requests=2.32.3=py39haa95532_1
scipy=1.10.1=py39hfbf2dce_0
setuptools=75.1.0=py39haa95532_0
tensorboard=2.10.1=pypi_0
tensorflow=2.10.0=mkl_py39ha510bab_0
tensorflow-gpu=2.10.0=pypi_0
tqdm=4.67.1=pypi_0
uvicorn=0.34.0=pypi_0
wcwidth=0.2.13=pypi_0
urllib3=2.2.3=py39haa95532_0